feat: configuring logger
