+++ server/core/log/log_configs.py
@@ -1,13 +1,16 @@
-import logging
 from core.settings import conf
 from core.errorfactory import LogsNotEnabled
+from importlib import import_module
 
 
 class Log:
+    # TODO: Find a better way to do this (if required)
+    #! init class only if configurations allow it
+    #! Ristrict all logging if logger is disabled
     def __new__(cls, *args, **kwargs):
         if conf["logs"]:
             return super().__new__(cls, *args, **kwargs)
         raise LogsNotEnabled
 
     def __init__(self, *args, **kwargs):
-        self._is_enabled = True
+        self.logger = import_module("logging")
